requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
xrpl-py==2.6.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
aiosqlite==0.20.0
